ai_modeling/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
beautifulsoup4
pydantic
python-multipart
pandas
openpyxl
langgraph
langchain
sentence-transformers
scikit-learn
pandas
